Migrate samples back to JS.
