+++ ai-assistant/index.html
@@ -150,6 +150,6 @@ For licensing, see LICENSE.md or http://ckeditor.com/license
 		</div>
 
 		<script src="https://cdn.ckbox.io/CKBox/2.4.0/ckbox.js"></script>
-		<script type="module" src="index.ts"></script>
+		<script type="module" src="index.js"></script>
 	</body>
 </html>
