+++ headless/index.html
@@ -20,6 +20,6 @@ For licensing, see LICENSE.md or http://ckeditor.com/license
 			<div id="cke5-headless"></div>
 		</div>
 
-		<script type="module" src="index.tsx"></script>
+		<script type="module" src="index.jsx"></script>
 	</body>
 </html>
