+++ headless/index.jsx
@@ -6,7 +6,7 @@
 import React, { useState, useEffect, StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { CKEditor } from '@ckeditor/ckeditor5-react';
-import HeadlessEditor from './index.ts';
+import HeadlessEditor from './index.js';
 
 function App() {
 	const [editor, setEditor] = useState(null);
@@ -59,7 +59,7 @@ function App() {
 				onReady={(editor) => {
 					setEditor(editor);
 
-					(window as any).editor = editor;
+					window.editor = editor;
 				}}
 			/>
 		</div>
@@ -140,19 +140,12 @@ function EditorToolbar({ editor }) {
 	);
 }
 
-type EditorToolbarButtonProps = {
-	label: string;
-	editor: HeadlessEditor;
-	commandName: string;
-	commandValue?: string;
-};
-
 function EditorToolbarButton({
 	label,
 	editor,
 	commandName,
 	commandValue,
-}: EditorToolbarButtonProps) {
+}) {
 	const command = editor ? editor.commands.get(commandName) : null;
 	const [isOn, setIsOn] = useState(false);
 	const [isEnabled, setIsEnabled] = useState(true);
@@ -163,15 +156,15 @@ function EditorToolbarButton({
 		}
 
 		function handleValueChange() {
-			if (typeof command!.value === 'boolean') {
-				setIsOn(!!command!.value);
+			if (typeof command.value === 'boolean') {
+				setIsOn(!!command.value);
 			} else {
-				setIsOn(commandValue === command!.value);
+				setIsOn(commandValue === command.value);
 			}
 		}
 
 		function handleIsEnabledChange() {
-			setIsEnabled(command!.isEnabled);
+			setIsEnabled(command.isEnabled);
 		}
 
 		command.on('change:value', handleValueChange);
