+++ headless/package.json
@@ -15,9 +15,9 @@
 		"preview": "vite preview"
 	},
 	"devDependencies": {
-		"@ckeditor/ckeditor5-react": "5.0.5",
-		"ckeditor5": "nightly",
-		"ckeditor5-premium-features": "nightly",
+		"@ckeditor/ckeditor5-react": "alpha",
+		"ckeditor5": "alpha",
+		"ckeditor5-premium-features": "alpha",
 		"vite": "^5.0.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0"
