+++ markdown/index.html
@@ -62,6 +62,6 @@ You can also use Markdown to create various text blocks, such as:
 			</div>
 		</div>
 
-		<script type="module" src="index.ts"></script>
+		<script type="module" src="index.js"></script>
 	</body>
 </html>
