+++ mathtype/package.json
@@ -15,7 +15,7 @@
 		"preview": "vite preview"
 	},
 	"devDependencies": {
-		"ckeditor5": "nightly",
+		"ckeditor5": "alpha",
 		"@wiris/mathtype-ckeditor5": "8.7.1",
 		"vite": "^5.0.0"
 	}
