+++ mobile/index.js
@@ -40,7 +40,7 @@ import {
 import 'ckeditor5/ckeditor5.css';
 
 ClassicEditor.create(
-	document.querySelector('#cke5-mobile-demo') as HTMLElement,
+	document.querySelector('#cke5-mobile-demo'),
 	{
 		plugins: [
 			Alignment,
@@ -138,7 +138,7 @@ ClassicEditor.create(
 	}
 )
 .then((editor) => {
-	(window as any).editor = editor;
+	window.editor = editor;
 })
 .catch((error) => {
 	console.error(error.stack);
