+++ source-code-editing/index.html
@@ -103,6 +103,6 @@ For licensing, see LICENSE.md or http://ckeditor.com/license
 		</div>
 
 		<script src="https://cdn.ckbox.io/CKBox/2.0.0/ckbox.js"></script>
-		<script type="module" src="index.ts"></script>
+		<script type="module" src="index.js"></script>
 	</body>
 </html>
