+++ user-interface-balloon/index.html
@@ -78,6 +78,6 @@ For licensing, see LICENSE.md or http://ckeditor.com/license
 			</div>
 		</div>
 
-		<script type="module" src="index.ts"></script>
+		<script type="module" src="index.js"></script>
 	</body>
 </html>
