+++ user-interface-button-grouping/index.html
@@ -64,6 +64,6 @@ For licensing, see LICENSE.md or http://ckeditor.com/license
 				</p>
 			</div>
 		</div>
-		<script type="module" src="index.ts"></script>
+		<script type="module" src="index.js"></script>
 	</body>
 </html>
